requirements.txt
# Core LangChain dependencies
langchain
langchain-google-genai
langchain-chroma

# UI Framework
streamlit

# Environment management
python-dotenv